+++ src/game/QueryHandler.cpp
@@ -160,9 +160,9 @@ void WorldSession::HandleCreatureQueryOpcode(WorldPacket & recv_data)
             CreatureLocale const *cl = objmgr.GetCreatureLocale(entry);
             if (cl)
             {
-                if (cl->Name.size() > loc_idx && !cl->Name[loc_idx].empty())
+                if (cl->Name.size() > uint32(loc_idx) && !cl->Name[loc_idx].empty())
                     Name = cl->Name[loc_idx];
-                if (cl->SubName.size() > loc_idx && !cl->SubName[loc_idx].empty())
+                if (cl->SubName.size() > uint32(loc_idx) && !cl->SubName[loc_idx].empty())
                     SubName = cl->SubName[loc_idx];
             }
         }
@@ -224,9 +224,9 @@ void WorldSession::HandleGameObjectQueryOpcode(WorldPacket & recv_data)
             GameObjectLocale const *gl = objmgr.GetGameObjectLocale(entryID);
             if (gl)
             {
-                if (gl->Name.size() > loc_idx && !gl->Name[loc_idx].empty())
+                if (gl->Name.size() > uint32(loc_idx) && !gl->Name[loc_idx].empty())
                     Name = gl->Name[loc_idx];
-                if (gl->CastBarCaption.size() > loc_idx && !gl->CastBarCaption[loc_idx].empty())
+                if (gl->CastBarCaption.size() > uint32(loc_idx) && !gl->CastBarCaption[loc_idx].empty())
                     CastBarCaption = gl->CastBarCaption[loc_idx];
             }
         }
@@ -277,7 +277,7 @@ void WorldSession::HandleCorpseQueryOpcode(WorldPacket & /*recv_data*/)
     int32 corpsemapid = mapid;
 
     // if corpse at different map
-    if (mapid != _player->GetMapId())
+    if (uint32(mapid) != _player->GetMapId())
     {
         // search entrance map for proper show entrance
         if (MapEntry const* corpseMapEntry = sMapStore.LookupEntry(mapid))
@@ -355,9 +355,9 @@ void WorldSession::HandleNpcTextQueryOpcode(WorldPacket & recv_data)
             {
                 for (int i = 0; i < 8; ++i)
                 {
-                    if (nl->Text_0[i].size() > loc_idx && !nl->Text_0[i][loc_idx].empty())
+                    if (nl->Text_0[i].size() > uint32(loc_idx) && !nl->Text_0[i][loc_idx].empty())
                         Text_0[i]=nl->Text_0[i][loc_idx];
-                    if (nl->Text_1[i].size() > loc_idx && !nl->Text_1[i][loc_idx].empty())
+                    if (nl->Text_1[i].size() > uint32(loc_idx) && !nl->Text_1[i][loc_idx].empty())
                         Text_1[i]=nl->Text_1[i][loc_idx];
                 }
             }
@@ -424,7 +424,7 @@ void WorldSession::HandlePageQueryOpcode(WorldPacket & recv_data)
                 PageTextLocale const *pl = objmgr.GetPageTextLocale(pageID);
                 if (pl)
                 {
-                    if (pl->Text.size() > loc_idx && !pl->Text[loc_idx].empty())
+                    if (pl->Text.size() > uint32(loc_idx) && !pl->Text[loc_idx].empty())
                         Text = pl->Text[loc_idx];
                 }
             }
